Added some benchmarks to the mux package.
These benchmarks indicate that multiplexed message sending takes about
twice as long as raw message sending over an io.Pipe and about the same
amount of time over an os.Pipe. Over an os.Pipe, we can send about
30,000 50kB messages per second, which works out to 1.5 GBps. In
practice, the receiving end of the os.Pipe is going to be
bandwidth-limited to a few MBps, so once that's saturated, we'll hardly
see any overhead from the multiplexing.

We could try to add an optimization to in-memory multiplexing to disable
buffering when going over an io.Pipe, as we will for in-memory
endpoints, but I think we'll also be limited in terms of
message-processing rate on the rsync side (every operation is doing a
disk write or disk read/write), so I think there's no point.

For reference, doing something like

yes | pv | cat > /dev/null

yields speeds of 0.68 GBps. I think the slower speed is due to the data
going across-processes (and 3 processes instead of 1). But this is a
reasonable upper-bound for what we'd ever be able to do across SSH.