Added reusable rsync.Engine type.
This commit refactors the rsync code to be a re-usable engine type
instead of free functions. This allows us to re-use buffers.

The rsync package builds in this commit, but it breaks interfaces used
by the rest of Mutagen. I'm committing it because I want to revert to
non-Protocol Buffers structures for encoding, but want a record of the
code working with Protocol Buffers structures.

I think we'll switch back to using encoding/gob for rsync. It may be
slower, but we'll almost certainly still be bandwidth-limited, and I
think the allocations required by moving to Protocol Buffers are too
much.