Made build work again, partially refactored controller/agent.
This commit adds some one-shot rsync functions that can be used until we
can get the endpoint refactored to re-use rsync engines.

It also removes our multiplexing package - this is probably not the
right way to go.

Finally, it adds a few cleanups to agent and controller code, but much
more awaits.