Massive refactor to avoid Muxado and use static multiplexing.
This is the first part of an effort to remove Muxado from our
dependencies and to use a static multiplexing scheme. Unfortunately, the
tight coupling between controller and endpoint required making all of
these changes at the same time, but hopefully the coupling will be able
to relax a bit now.

The main change is that rsync has been modified to be a streaming,
pipelined protocol rather than use the RPC package with concurrent
requests to achieve pipelining.

We've also switched to a fixed multiplexing scheme that doesn't require
dynamic channel forwarding.

There are also a large number of optimizations to rsync transfers.

Scan and watch have also been separated. Watch is currently disabled and
replaced with a timer, but this will be fixed shortly.

CPU overhead has been significantly reduced (likely due to the lack of
dynamic multiplexing and less garbage being generated). Transfer time
has also been reduced by about 25% and takes about 1.25 times as long as
uncompressed rsync (1.5 times as long as compressed rsync).

Muxado and the RPC package still exist for the daemon, but that will be
fixed shortly.