Changed rsync protocol to just send batched requests.
This brings our implementation up to parity with rsync in terms of
transfer speed and CPU usage (i.e. significantly faster with
significantly less CPU usage).

The reason for the previous design was bad premature optimization. We
assumed that the overhead for closing and re-opening files would be bad,
but it is not. We also assumed that we might use a large quantity of
memory if receiving many signatures, but this case is very unlikely
since we only send signatures for every file if every file has changed
(which they likely haven't), and on new file transfers, we mostly send
nil signatures.

Anyway, this is WAY simpler and WAY better.