Modified controller scan function to allocate rsync engine.
We were previously using a cached engine that was shared between scan
calls. But we're going to do some refactoring that'll require the scans
to run concurrently to allow for cancellation, so we can't share any
more. We could allocate as a member of the controller, but that leaves
memory hanging for the life of the controller, and we're going to modify
rsync engines to lazily allocate anyway, so the cost will be very small.
The scan function already allocates so much memory that the rsync engine
will be a drop in the bucket.