Changed rsync client state monitoring to use state package.
This commit changes the rsync client state monitoring to use the state
package for tracking instead of a callback. This necessitated a few
changes to the state package, namely the ability to unblock tracking.
Corresponding changes were also necessary in the session package, but
these actually enhanced functionality.

This improves performance substantially, because each rsync receive is
no longer blocking waiting for the status update to send. The exact
difference is hard to measure, but based on CPU profiling and timing,
I'd estimate 5-10%.

This does necessitate another Goroutine in the endpoint, but this
actually makes the design more consistent.

One thing that's surprising is that the condition variable broadcast in
state tracking is quite expensive. It might be worth seeing if we can
optimize that.