Made staging interruptable.
This is necessary since staging can take a long time, and if someone
wants to pause or terminate a session, there's no need to wait around
for it to complete.

We may also want to extend this to scanning at some point, but it's
relatively quick. We'll never want to extend it to transition, which
shouldn't be interrupted.