Modified staging to skip staging already staged files.
This is a hugely important bug fix and the whole reason that we created
the staging directory to begin with. Now the staging method will first
check if files have already been staged and avoid restaging them if so.
This of course makes use of all the checksum infrastructure built into
staging.