Added create parameter to filesystem.Mutagen.
The filesystem.Mutagen was designed to always invoke os.MkdirAll when
computing a path, which is useful in many cases, but profiling very hot
in transition (where it wasn't necessary since transition paths should
already completely exist or fail). This commit adds a parameter that
controls that behavior and updates existing code to use it as necessary.