Updated endpoint to use native filesystem watching when possible.
This implementation is not ideal, but it should be fairly robust. Native
watching is only performed on systems that provide native recursive
watching (i.e. Windows and macOS). Other systems that use inotify,
kqueue, et al. are vulnerable to file descriptor exhaustion on even
modestly sized roots.

Even on Windows and macOS, however, there is a high degree of
variability in terms of robustness and behavior of filesystem
monitoring. On Windows, it's not possible to watch a file or to monitor
for deletion of a watch root. Deletion of a watch root on Windows also
causes a number of problems with the Go package we're using for watching
(github.com/rjeczalik/notify), which is unfortunately the best there is.
Even if we could detect watch root deletions reliabily and recover from
them (which we can't), it would be difficult to determine which watch
errors are terminal and which aren't.

In order to work around these issues, systems supporting native watching
will simply watch the user's home directory and filter out events that
don't occur at or below the watch root. Because the home directory is
persistent (or at least we rely on its persistence for other features),
we can be sure that a watch at that location will be stable. This does
mean that native watching can only be done for paths which are
descendants of the home directory, but this should at least handle the
majority case, especially on Windows and macOS.

For all other systems and in all other cases, we fall back to filesystem
polling, which, while not efficient, is more scalable.

In the future, it might be worth trying to partially support systems
without native recursive watching via some most-recently-updated
mechanism, but this will be very complex.