Updated vendored kardianos/osext dependency.
It seems that someone else finally caught onto the fact that the OpenBSD
mechanism was broken in the os package's implementation:

https://github.com/kardianos/osext/issues/22
https://github.com/golang/go/issues/19453

A fix will be available in Go 1.9, and in the mean time the osext
package has been updated to use its old method. Once 1.9 rolls out,
we'll get rid of the osext dependency.