Fixed two related issues with filesystem polling.
This commit fixes two major bugs with filesystem polling:

First, if an error occurred during polling, all polling was halted and
the Goroutine would wait indefinitely because the timer wasn't reset.

Second, if the polling root didn't exist (which is a valid state that
needs to be able to trigger changes, e.g. on deletion), then an error
was returned from the poll method. This commit fixes polling to be smart
enough to understand the root not existing and determine changes in that
case. All other errors are still treated as being due to concurrent
modifications.