Devirtualized and renamed message streams.
This commit removes the MessageStream interface and replaces it with a
concrete type: Stream. There was no need for this interface. It was
originally intended to mask compressed vs. non-compressed message
streams, but they ended up being implemented by the same type (with a
completely negligible performance hit to non-compressed streams), so it
was better to just expose this type directly. It also made sense to
remove the word "Message" from the name since the package name already
indicates the type of stream.