Restructured Travis build matrix and enabled coverage reports.
This commit is the first attempt to restructure the Travis CI build
matrix, making it simpler to specify which build configuration is
responsible for release build and coverage profile uploading.

This commit also attempts to implement merged coverage profiles. It
will probably require a few revisions to get this right.