Removed coverage report generation, merging, and uploading.
I don't really like Coveralls.io, and it's asking for a few too many
permissions. Additionally, it's better to evaluate code coverage on all
platforms rather than in just one configuration. Since we're keeping the
command line coverage statistics, we can look at coverage statistics for
a particular build that we're interested in.