Implemented code coverage using codecov.io.
This commit implements code coverage monitoring using codecov.io. It may
take a few revisions to get functional. It adds a custom script for
invoking tests and generating coverage profiles. It then adds a coverage
profile merging step and a step to upload the resulting report to
codecov.io. The upload is performed with the Bash uploader on macOS and
Linux and the Python uploader on Windows. codecov.io merges coverage
reports, so this should take us closer to understanding cross-platform
coverage in the way that we want.

This commit also reverts the explicit configuration matrix added to
.travis.yml because it was more trouble than it was worth.