Modified CI testing script to invoke gocovmerge.
Windows' cmd.exe doesn't support wildcards, so we need an alternative
mechanism for passing coverage profiles to gocovmerge. Since the
ci_testing.go script knows the location of these files, it makes sense
for it to just invoke gocovmerge directly, and it also gives us a chance
to use a temporary directory to house these files.