Added code coverage badge to README.md and restructured badge table.
This will probably be a bit of an embarrassment for a while since code
coverage is low, but it's also an essential metric that we should strive
to increase.

Since we needed to restructure the badge table anyway, I've taken the
liberty of changing OS X references to macOS references.