Removed custom copy mechanism from multiplex package.
I'm not sure what I was thinking, but io.LimitedReader exposes all of
its internals, so it's perfectly fine to just reset it and reuse it.
This saves the need for allocations on every read. We can also detect
early EOF behavior by just looking at the number of bytes transferred,
so there's not really any issue there either. There's no performance
gain to be had here - maybe just a microscopic reduction in the amount
of code loaded into memory.