Bumped required version to Go 1.9.
This is necessary to fix two issues.

First, prior to Go 1.9, an os.Lstat call on Windows could set both the
os.ModeDir and os.ModeSymlink bits for a directory symlink. This has
been fixed via a few CLs:

https://go-review.googlesource.com/41830
https://go-review.googlesource.com/41834

The logic we're using is already sound under the assumption that the
os.ModeDir bit is not set for symlinks, which should always be the case.

This will also be required to finally remove our dependency on the osext
package (which was required due to a flakey implementation of
os.Executable on OpenBSD in Go 1.8).