Switched to using os.Executable for executable path detection.
This has been available since Go 1.8, but it's been flakey (read:
effectively broken due to the lack of /proc) on OpenBSD. This was fixed
in Go 1.9, so we can finally use it.

This also negates the need for the github.com/kardianos/osext package,
which I've removed from vendoring and license information.