Enabled OpenSSH ControlMaster for connection re-use.
This will allow re-use and sharing of SSH connections opened to remote
hosts. This will work across both ssh and scp commands. It should reduce
the number of manual password entries required to connect, install
agents, and start sessions.

Connections are stored within the ~/.mutagen directory (in the
connections subdirectory). This is only supported on POSIX systems since
Windows doesn't support UNIX domain sockets, which are necessary for
this feature to be implemented by OpenSSH (though there is some
indication that the PowerShell port may eventually used Windows Named
Pipes to support this).