Modified all CI environments to test for GOARCH=386.
AppVeyor was previously testing using the i386 Go installation, but this
didn't port well to Travis, where we at least want to test this
configuration for Linux. This commit modifies our ci_testing.go script
with a flag that allows testing for GOARCH=386, as well as a flag for
disabling coverage output (since we'll just take the amd64 cover reports
for now). All CI configurations have been updated to use this new
feature, meaning that Windows, macOS, and Linux all test with
GOARCH=386. This also cleans up the AppVeyor configuration as a result.

This commit also adds a CopyCurrent utility function to the environment
package since this was becoming a recurring operation. Other code using
copies of the current environment have been updated to use this
function.