Added slight optimization to reconciliation algorithm.
Rather than computing both ancestor->alpha and ancestor->beta diffs up
front, we can compute one, see if it allows us a fast return, and then
compute the other diff if not and check that. This could potential save
us at least one large DFS traversal per synchronization cycle.

This is slightly more advantageous if we assume that alpha is more
likely to be unmodified, but since our sync is completely symmetric, we
don't really have a way to assess whether or not that's the case. We
might make a recommendation to users to make the alpha the
less-frequently modified side, but I think we should just try for the
optimization if we can get it and if not then deal with the microscopic
cost of computing an additional diff. If we do decide to "recommend" an
ordering, we can rearrange these checks to allow the optimization to
kick-in for beta being unmodified if that's more "natural."

While this whole thing is sort of a micro-optimization, it doesn't make
the code structure any less clear, and I think it's worth it for very
large trees.