Added version to release archive names.
This commit modifies the build scripts to add the Mutagen version to the
release archive names. It also modifies the deployment configuration for
Travis to use globs, removing the need to manually specify bundles. This
latter change will result in .tar.gz archives also being uploaded for
Windows, but that's fine since some people may be using POSIX
environments where those archives are easier to open.