Explicitly bumped Travis Go version to 1.9.2.
For some reason, using 1.9.x results in getting Go 1.9, even though
1.9.2 is out. Since we need 1.9.2 to fix the Windows standard input
reading bug, we explicitly request it.

We don't have corresponding control on AppVeyor, but we don't use that
for release builds, and at the moment we don't have any tests for
standard input reading on Windows that might lock up, so we can just
take whatever version of Go 1.9 we get there, at least for the time
being.