Removed comment about relocating encoding routines.
This may still eventually happen, but I think we're more likely to bring
in a Mutagen configuration file before that, probably using TOML, so I
want to leave this package in place for eventual expansion. Nothing
about the encoding package's functionality is hyper-specific to the
session package.

The session package is also already overloaded with functionality, so
I'm actually a bit averse to moving more code into it.