Added documentation on justification for storing Go mode bits.
The truth is that this design, where we store Go-specific mode bits, was
more of an accident than a conscious decision. We did know that the bits
were technically "portable," but honestly didn't know about the
cleaner/simpler enumeration-based design that we could have gone with.

In any case, after thinking about it, there is some value in this
accident, and thus I don't want to rush into changing it. It's still
technically a valid design, if not the best, and it may allow for some
later developments that would actually be harder with the "cleaner"
design.

This commit documents the logic in our thinking and the reason for not
rushing into "fixing" the design.