Removed comment about racy Git solution.
The "racy Git" solution doesn't really apply to us because it is
predicated on modifications happening serially, not concurrently. In
fact, a lot of Git's design is predicated on this assumption (e.g.
stat'ing and hashing are not a single atomic operation, so technically
the hash in Git's index might not even match the timestamp associated
with that entry if the file is modified between stat'ing and hashing or
concurrently with hashing). We have no control over concurrent
modifications, and thus checking timestamps against the cache file
timestamp really has no meaning or value to us.

In any case, it's less of a problem these days anyway with
high-resolution modification timestamps on ext4 (1 ns), APFS (1 ns), and
NTFS (100 ns).

If changes really aren't detected, the user can always just manually
touch the file to trigger a full rehashing.