Added codecov.io YAML configuration to ignore generated code.
This commit adds a YAML configuration file for codecov.io that instructs
it to ignore generated Protocol Buffers code when computing code
coverage. It's possible to get gogo/protobuf to generate tests, but it
requires enabling a number of extensions that just bloat code. Since
testing it really won't make much difference, we just ignore it.