Fixed issue with vestigial environment variables on Windows.
Windows has a lot of weird environment variables that it uses to
maintain backwards compatibility (or some other black magic - they're
undocumented). They start with '=' (so they technically have empty
names, but not really - the variable name just starts with '=' and
GetEnvironmentVariable treats them specially), they have weird names
(like "=::"), and they only really show up when using
GetEnvironmentStrings (which is what Go's runtime uses to populate the
result of os.Environ).

So we just ignore them, because they'll never be useful to us. This
works for POSIX too, where they are effectively useless and possibly
even invalid (though I can't find anything in the specification that
supports that conclusion).