Restructured housekeeping and added it to daemon.
This commit exports the session housekeeping functions and just has the
agent invoke them instead of doing it inside ServeEndpoint. The daemon
has also been updated to perform agent and session housekeeping, since
ServeEndpoint won't be responsible for doing it locally now.

The reason for this is mostly just symmetry and consolidation of
housekeeping.