Massively refactored rsync transmission design.
This commit removes the rsync.Client/Serve infrastructure in favor of a
far more flexible and efficient Transmit/Receiver design. The
rsync.Transmit function is very similar to rsync.Serve, but it knows
nothing about the network or encoding, instead pushing operations to an
rsync.Receiver type that handles their application. This necessitated
changes to the way that rsync.Engine handles Patch operations, making it
push rather than pull based, and thus far more flexible.

This commit breaks the build, but I wanted it isolated in the commit
history.