Removed error from filesystem.Watch.
The only error we returned was on cancellation, which isn't very useful.
In the event of multiple poll failures, we may want to fall back to a
timer that fires every N seconds to cause a rescan, but we can implement
that later. If a root can't be polled, there are likely bigger issues.