Modified rsync.Engine to use optimal block size if possible.
The rsync.Engine.Signature method technically doesn't require an
io.ReadSeeker, only an io.Reader (even though 99% of the time it will
receive a base that can seek since that's necessary for
rsync.Engine.Patch), so this commit fixes that interface requirement. If
the base does happen to implement io.Seeker, the engine will now use the
optimal block size for that base if no block size is specified. If the
base doesn't implement io.Seeker, or there's an error determining the
optimal block size, then the default block size is used.