Got rid of Snappy compression in favor of SSH compression.
Unfortunately the Go compression story is not ideal. The DEFLATE
implementation is slow. The Snappy implementation is fast, but its
compression ratio is significantly worse than DEFLATE's, which is why
rsync has a huge edge over Mutagen (it has DEFLATE and it's 2-3x faster
than Go's).

This commit removes the use of Snappy and instead just piggybacks on
SSH's built-in compression, which will be DEFLATE-based and much better
optimized for the platform.

We can revisit this as Go's compiler and compression implementations
improve.

Removing Snappy also saves us a tiny bit of binary size.

Finally, since we use SCP compression as well, agent binaries should
transfer faster.