Re-enabled watch coalescing - correctly this time.
This will be necessary during transitions on systems with recursive
watching, otherwise we'll have a huge number of events going across
the wire.

It turns out that we weren't actually properly draining the coalescing
timer in the last implementation - you have to do a non-blocking drain
if you don't know that it was running.