Modified controller to completely drain pollers before synchronizing.
It makes sense to completely drain the pollers to avoid another
synchronization pass due to events that had already occurred before
starting the previous pass. This is primarily useful for sessions where
changes are coming in frequently from both endpoints, and even then it
really only saves you one pass, but it is the optimal behavior.