Revert "Switched to syscall.Rename instead of os.Rename to avoid os.Lstat."
This reverts commit e44ca1b6d967a9d512db29ac5084dca97ba8da94.

I'm not convinced that this optimization is worth the associated
technical debt in terms of needing to be mindful the length of paths
within the staging directory. In fact, it may just end up being
incorrect on Windows, because we really have no good idea of the length
of %USERPROFILE% and we're losing long-path support by calling directly
into the syscall package.

The main reason this was profiling so hot to begin with is because it
was on a macOS system with an encrypted APFS partition. It's
well-documented online that the performance of encrypted APFS
partitions isn't excellent at the moment, but this is something I expect
Apple will optimize over time, probably sooner rather than later.