Switched endpoints to create rsync engines dynamically.
These are not expensive to allocate, and it's just more bookkeeping to
store them in the endpoint structure. It's also better not to leave them
lingering since their internal buffers never shrink. They probably
aren't a huge space driver, but they should be thought of in the same
sense as copy buffers: reusable in the short term, but not worth keeping
around for the long term.