Added permission preservation.
Mutagen still doesn't support permission propagation and probably never
will. E.g., you might have files marked 0666 on your laptop that you
want set to 0600 on a remote server with many users. But, if you have a
file set to 0666 on your laptop and edit it on a remote server and sync
it over, you want to keep those permissions in place. This commit adds
that functionality by modifying providers to accept a base file mode to
use when computing permissions, which is fortunately readily available
by our stat checks. This propagation only occurs for files being
updated, not newly created files, which still inherit a base permission
of 0600 (or 0700 if executable). It also doesn't apply to directories or
symlinks, but doesn't need to because those aren't replaced in the same
way (the directory itself is not recreated if its contents changed, so
permissions are already "preserved" in that case). Executability is also
added on a user-only basis and stripped on an everyone basis when added
or removed, respectively, but it will be preserved so long as it doens't
changed.