Implemented root deletion safety mechanism.
This commit adds a best-effort safety mechanism to avoid propagating
root deletions. Basically, if a session detects root deletion, it
immediately halts into a mode that registers below connectivity (but
above complete disconnection). In order to resume the session, the user
can manually delete the root on the other side where the deletion would
have been propagated and then resume the session (at which point it will
reconnect).

For files, this detection should work perfectly.

For directories, there's the problem that a scan may occur at a point
where the root is being removed but still partially exists. In this
case, the safety mechanism won't kick in. I've modified the scan
infrastructure to completely abort if it detects any concurrent
modifications, but this is not perfect. It's possible that a scan could
just see a large portion of root deleted without detecting that it's
part of a larger ongoing deletion. We sort of just have to live with
this. I've had ideas about using some heuristic, e.g. % of files deleted
within a single propagation, to halt in the same way, but it's difficult
to tune that. We could certainly make it a parameter in the future. We
could also refuse to propagate deletions, but that would make a lot of
legitimate work very tedious (since the user would have to manually
propagate the deletion).

In any case, this should handle may important cases, and it's better to
have some safety check in place than none.