Disabled symlinks as synchronization roots.
This is a breaking but necessary change. There's not really any way to
support this that doesn't involve horrendous (and probably not 100%
correct) hacks. The main problems are (a) watches can't catch changes to
symlinks and (b) symlink roots that point to non-existent files will
fail with "already exist" errors when propagating the root from the
other side and (c) symlink roots that DO exist will have the symlink
itself wiped out when changes are propagated from the other side.

Existing sessions with symlink roots will just enter a "waiting for
rescan" failure because scans will consistently fail. No data will be
lost.