Replaced underlying UUID package.
This commit replaces the github.com/satori/go.uuid package with the
github.com/google/uuid package. Vendoring has been updated, as has the
license text.

The go.uuid API had recently been broken, and in investigating this it
was revealed that the V4 (random) UUID implementation in go.uuid was
broken and had been for some time. The Google implementation seems to be
correct. The Google implementation's API is not guaranteed to be stable
just yet, but it seems highly unlikely that any breaking changes will
result in significant code changes on our side (we might just need to
change the name of NewRandom). If we continue to see breaking changes in
the UUID landscape, it may be worth paving over this with our own
internal wrapper.