Added validation for Protocol Buffers loaded from disk.
This commit adds various validators (in addition to those that already
exist) for ensuring valid and sane Protocol Buffers have been loaded
from disk. Once in memory, we assume the invariants that we enforce
here, and even though we essentially control all files in ~/.mutagen, we
don't want to crash the whole daemon if some untrusted or invalid or
corrupted message was saved to disk.