Performed a massive code relocation and updated CI configuration.
We're going to be adding a lot of new non-Go code "stuff" to the
repository in the next few weeks, and go test doesn't handle this
particularly well. The source tree was also getting a bit hard to read
at its root due to all the packages. This commit combines all non-Main
packages into a pkg/ directory and fixes all import paths accordingly.

This also allows us to ditch the ci_testing.go script and vastly
simplify CI configuration. With Go 1.10 now supporting multi-package
coverage profiles, things are really clean.

We've had to temporarily disable 386 testing on Windows (due more to
complexity reasons than technical reasons), but (a) 386 code is already
tested on Linux and macOS, (b) there are already many other
architectures that we don't test, and (c) we'll be trying to fix (b) in
the next few weeks (perhaps with QEMU) and can address this then.