Removed ControlMaster usage.
This commit removes usage of SSH's ControlMaster functionality. It
doesn't preclude its usage (e.g. via configuration in ~/.ssh/config),
but it stops Mutagen from assuming that it should be used. We've put in
a lot of effort to make Mutagen respect OpenSSH configuration, and it
doesn't make sense to fight the user's settings here.

The only place that this will cause a noticable difference in the UI is
for people using password-based authentication on POSIX systems
(ControlMaster was never supported on Windows). People using automatic
methods of authentication won't notice a difference, except perhaps in
connection time. However, users can always manually re-enable
ControlMaster (and Mutagen will respect this), which is the way that
this should have been done to begin with. We will add this to the
documentation.

Automatic setting of ConnectionTimeout and compression usage may fly in
the face of the above logic, but I've left them for now because I can't
think of any scenario where they don't make sense for Mutagen. This is a
bit of a heuristic/judgement call, but I think it's a sane one.

If we do decide to add this support back in the future for some reason
(which I would advise against), we should definitely think through the
folder structure a bit better. If we eventually want to support multiple
transports, then ~/.mutagen/connections is not the best place to store
these sockets.

Fixes #25.