Reverted licenses to their standard formatting.
This commit reverts our own license as well as third-party license
notifications to their standard formatting. It was nice to have these
formatted for easier reading, but it's going to be a painful experience
to have to manually reflow the text for every license in the future.
Let's just copy and paste them in.

This commit also renames out license file to be non-Markdown, better
adhering to common practice.

We did have to reformat the npipe license because the original text has
had the line breaks removed, but we've gone ahead and used the standard
MIT license formatting to accomplish this.