Modified rpc package to run on top of yamux.
This allows us to perform streaming RPC without the need for one
connection per call and without the need for something as heavy as gRPC.
It also cleans up the daemon IPC code and makes it a bit more canonical.

We've been waiting to do this for a while, but there were some issues
with yamux where the buffer for a stream would grow indefinitely. These
were recently fixed. yamux may eventually prove useful in simplifying
the remote endpoint implementation.

This commit also adds a vendored copy of yamux as a submodule and adds
the appropriate license information to the legal.go file.