Polished daemon start/stop for use with launchd.
When using a launchd-registered daemon, we need to use the launchctl
command to control its lifecycle, so this commit adds a few hooks to
make that possible.

We've also disallowed changing daemon registration status on macOS while
the daemon is running, because the start/stop mechanism is determined by
the daemon registration status, and we don't want to try to run into
conflicting daemon lifecycle management methods (e.g. trying to stop a
normally launched daemon using launchctl, or having launchd try to
restart the daemon over and over again while a normally launched daemon
owns the lock).