Added bash completion script generation.
This commit adds a (hidden) flag to the mutagen command to generate a
bash completion script. This will be used as part of the Homebrew
installation script (and perhaps other installation mechanisms) to
provide bash completion.

Performing autocompletion of remote URLs is going to be tricky. I may
punt that to a later release. In any case, the infrastructure to do this
is present.