Switched to using gRPC for daemon RPC.
This commit switches over to gRPC for the daemon's RPC needs. This is a
MASSIVE commit due to the fact that the existing RPC mechanism was
deeply ingrained in a number of packages and that gRPC requires Protocol
Buffers versions of any data structures it has to deal with. Hopefully
the restructuring of the services will make any future RPC changes much
less invasive.

The transition is relatively straightforward. Rather than try to convert
all packages types to Protocol Buffers types, a thin set of message
types has been created in the session package to wrap any sync/rsync
package types for gRPC. This is actually better, because it allows us to
transmit shallow sync.Entry copies for conflicts. It does have the
downside of making certain state components nullable, but we can deal
with this through some minor validation.

In order to avoid a gRPC dependency in the agents, gRPC Protocol Buffers
definitions, generated code, and associated service implementations have
been put into service subpackages of their respective packages. This has
the nice side-effect of unifying all services into what Protocol Buffers
considers to be the "service" package.

This commit also removes the existing rpc package as well as our
vendored copy of yamux and its associated license.

This commit adds vendored dependencies for gRPC and its dependencies, as
well as their associated licenses.

Tremendous effort has also been put into cleaning up the license file to
be far more readable.