Added utility function to strip away RPC error wrappers.
When presenting error messages to users, it's useful to strip away the
gRPC error layer that sits around the underlying error message. This
commit adds a function, peelAwayRPCErrorLayer, that attempts to strip
away any surrounding gRPC error layer. This is only used when invoking
methods and communicating on streams - it's not used for dialing and the
like.