Tried to improve SSH encoding situation.
We now explicitly request a C locale when using SSH. Whether or not the
server respects this is outside our control, but we do validate that
whatever it sends back is UTF-8, and if it's not then, well... we'll
cross that bridge once users start coming to it.

The place where I think this may be an issue is the PowerShell OpenSSH
implementation. It's really unclear what the story is there with respect
to encoding support. We'll have to wait to see what happens as it moves
out of its beta state.