Fixed issue where code relied on directory contents being sorted.
There was one code instance that was broken by
964dffbae0950ebd3977074864311c9f5dd1e289. It was attempting to use a
binary search to locate entries. It has been updated to just use a
linear search. It's not really justified to revert
964dffbae0950ebd3977074864311c9f5dd1e289, because (a) the sorting
behavior was only relied upon in this one place (and paid for
everywhere), and the linear search is only O(n) vs O(n*log(n)) for the
sort. If we start relying on the sorting behavior again, we can revisit
this issue, but I seriously doubt that will be the case.