Switched mutagen-agent to use Cobra and include legal/version commands.
This commit modifies mutagen-agent to use Cobra and also include legal
and version information. This comes at the cost of around 700 kB of
binary size on macOS, but this compresses to a 300 kB difference in the
resulting binary archive (which for 25 binaries would yield an increase
of 7.5 MB). I'm willing to live with this for now, especially if we
start looking to expand the functionality of the agent without breaking
the old hand-written command dispatch code.