Fixed Chmod invocation to be path-based in RenameFileAtomic.
os.File-based Chmod operations are not supported on Windows - only
path-based Chmod. This commit modified RenameFileAtomic to use the
latter to avoid the EWINDOWS error that was coming up.

As far as restricting the mode bits when calling Chmod, it turns out
that this isn't necessary, because the Go Chmod implementations do this
by default. Thus we can still use FileInfo.Mode results to "transfer"
permissions between files.